+++ src/VecSim/CMakeLists.txt
@@ -15,9 +15,9 @@ add_subdirectory(spaces)
 set(CMAKE_CXX_FLAGS "${CMAKE_CXX_FLAGS} -Werror -Wall")
 
 add_library(VectorSimilarity ${VECSIM_LIBTYPE}
-    algorithms/brute_force/brute_force_factory.cpp
-    algorithms/hnsw/hnsw_factory.cpp
-    algorithms/hnsw/tiered_factory.cpp
+    index_factories/brute_force_factory.cpp
+    index_factories/hnsw_factory.cpp
+    index_factories/tiered_factory.cpp
     algorithms/brute_force/vector_block.cpp
     algorithms/hnsw/visited_nodes_handler.cpp
     vec_sim.cpp
