+++ tests/benchmark/bm_updated_index.h
@@ -6,7 +6,7 @@
 #include "VecSim/vec_sim.h"
 #include "VecSim/query_results.h"
 #include "VecSim/utils/serializer.h"
-#include "VecSim/algorithms/hnsw/hnsw_factory.h"
+#include "VecSim/index_factories/hnsw_factory.h"
 #include "bm_common.h"
 
 /**************************************
